[FAB-3101] rename fabric shim java package
As part of cleanup towards a v1.0 release
moved:
  org.hyperledger.java
to:
  org.hyperledger.fabric.shim
