+++ core/chaincode/shim/java/src/main/java/org/hyperledger/fabric/shim/Chaincode.java
@@ -11,7 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-package org.hyperledger.java.shim;
+package org.hyperledger.fabric.shim;
 
 import org.hyperledger.fabric.protos.peer.ProposalResponsePackage.Response;
