+++ core/chaincode/shim/java/src/main/java/org/hyperledger/fabric/shim/fsm/CallbackKey.java
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package org.hyperledger.java.fsm;
+package org.hyperledger.fabric.shim.fsm;
 
 public class CallbackKey {
