+++ core/chaincode/shim/java/src/main/java/org/hyperledger/fabric/shim/fsm/EventDesc.java
@@ -1,4 +1,4 @@
-package org.hyperledger.java.fsm;
+package org.hyperledger.fabric.shim.fsm;
 
 /**
  * Represents an event when initializing the FSM.
