+++ core/chaincode/shim/java/src/main/java/org/hyperledger/fabric/shim/fsm/exceptions/NotInTransitionException.java
@@ -14,15 +14,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package org.hyperledger.java.fsm.exceptions;
+package org.hyperledger.fabric.shim.fsm.exceptions;
 
-public class UnknownEventException extends Exception {
+public class NotInTransitionException extends Exception {
 
-	public final String event;
-	
-	public UnknownEventException(String event) {
-		super("Event '" + event + "' does not exist");
-		this.event = event;
+	public NotInTransitionException() {
+		super("The transition is inappropriate"
+				+ " because there is no state change in progress");
 	}
 	
 }
