+++ core/chaincode/shim/java/src/main/java/org/hyperledger/fabric/shim/helper/Channel.java
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package org.hyperledger.java.helper;
+package org.hyperledger.fabric.shim.helper;
 
 import java.io.Closeable;
 import java.util.HashSet;
